requirements.txt
Flask==3.1.1
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.1.0
requests==2.32.3
pandas==2.2.3
matplotlib==3.10.3
plotly==6.1.2
python-dotenv==1.1.0
python-amazon-sp-api==1.9.33